# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Stock Game MVP API
  version: 0.1.0
  description: >-
    Internal training A-share simulation game API.
    Matching is server-authoritative with 5-minute batch matching and auction windows.
servers:
  - url: https://api.stock-game.internal
...
security:
  - bearerAuth: []
paths:
...
  /v1/seasons/{seasonId}/join:
    post:
      tags: [Seasons]
      summary: Join season and bind trading account (idempotent)
      parameters:
        - $ref: '#/components/parameters/SeasonId'
      responses:
        '200':
          description: Join result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JoinSeasonResponse'
        '404':
          description: Season not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /v1/seasons/{seasonId}/orders:
    post:
      tags: [Orders]
      summary: Place a limit order
      parameters:
        - $ref: '#/components/parameters/SeasonId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceOrderRequest'
      responses:
        '201':
          description: Order accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags: [Orders]
      summary: Query my orders
      parameters:
        - $ref: '#/components/parameters/SeasonId'
        - in: query
          name: status
          schema:
            $ref: '#/components/schemas/OrderStatus'
        - in: query
          name: tsCode
          schema:
            type: string
      responses:
        '200':
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
  /v1/seasons/{seasonId}/orders/{orderId}/cancel:
    post:
      tags: [Orders]
      summary: Cancel an active order
      parameters:
        - $ref: '#/components/parameters/SeasonId'
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Cancel rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SeasonId:
      in: path
      name: seasonId
      required: true
      schema:
        type: integer
        format: int64
  schemas:
...
    PlaceOrderRequest:
      type: object
      required: [clientOrderId, accountId, tsCode, side, limitPrice, quantity]
      properties:
        clientOrderId:
          type: string
          maxLength: 64
        accountId:
          type: integer
          format: int64
        tsCode:
          type: string
        side:
          $ref: '#/components/schemas/OrderSide'
        limitPrice:
          type: number
          minimum: 0.001
        quantity:
          type: integer
          minimum: 100
          description: Must be multiples of 100
    JoinSeasonResponse:
      type: object
      required: [seasonId, accountId, isNewJoin, initialCash, availableCash, frozenCash, realizedPnl]
      properties:
        seasonId:
          type: integer
          format: int64
        accountId:
          type: integer
          format: int64
        isNewJoin:
          type: boolean
        initialCash:
          type: number
        availableCash:
          type: number
        frozenCash:
          type: number
        realizedPnl:
          type: number
    Order:
      type: object
      required:
        [id, clientOrderId, tsCode, side, limitPrice, quantity, remainingQty, status, createdAt]
      properties:
        id:
          type: integer
          format: int64
        clientOrderId:
          type: string
        tsCode:
          type: string
        side:
          $ref: '#/components/schemas/OrderSide'
        limitPrice:
          type: number
        quantity:
          type: integer
        remainingQty:
          type: integer
        status:
          $ref: '#/components/schemas/OrderStatus'
        rejectCode:
          $ref: '#/components/schemas/RejectCode'
        rejectReason:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          $ref: '#/components/schemas/RejectCode'
        message:
          type: string
        traceId:
          type: string
    OrderSide:
      type: string
      enum: [buy, sell]
    OrderStatus:
      type: string
      enum: [pending, active, partially_filled, filled, canceled, rejected, expired]
...
    RejectCode:
      type: string
      enum:
        - LOT_SIZE_INVALID
        - LIMIT_PRICE_OUT_OF_BAND
        - SELL_T1_BLOCKED
        - INSUFFICIENT_SELLABLE_QTY
        - STOCK_HALTED
        - SEASON_NOT_TRADING
        - AUCTION_WINDOW_CLOSED
        - INSUFFICIENT_CASH
        - ORDER_NOT_FOUND
        - ORDER_NOT_CANCELABLE
        - ACCOUNT_NOT_FOUND
        - PERMISSION_DENIED
        - SEASON_NOT_FOUND
        - JOIN_SEASON_FAILED
...